separate code into MapPage.kt Profile.kt PointPage.kt and enlarge PointPage padding
