+++ app/build.gradle.kts
@@ -80,7 +80,8 @@ dependencies {
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
     implementation(libs.androidx.datastore.preferences.core.jvm)
-    implementation(libs.play.services.location)
+    implementation("com.google.android.gms:play-services-location:21.0.1")
+//    implementation(libs.play.services.location)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
