+++ app/src/main/java/com/example/lifelinealert/page/MapPage.kt
@@ -1,0 +1,96 @@
+package com.example.lifelinealert.page
+
+import android.Manifest
+import android.content.Intent
+import android.net.Uri
+import android.provider.Settings
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import com.google.accompanist.permissions.ExperimentalPermissionsApi
+import com.google.accompanist.permissions.isGranted
+import com.google.accompanist.permissions.rememberPermissionState
+import com.google.android.gms.maps.model.CameraPosition
+import com.google.android.gms.maps.model.LatLng
+import com.google.maps.android.compose.GoogleMap
+import com.google.maps.android.compose.MapProperties
+import com.google.maps.android.compose.MapUiSettings
+import com.google.maps.android.compose.Marker
+import com.google.maps.android.compose.MarkerState
+import com.google.maps.android.compose.rememberCameraPositionState
+
+//@Preview(showBackground = true)
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+fun MapPage() {
+    val cameraPermissionState =
+        rememberPermissionState(permission = Manifest.permission.ACCESS_FINE_LOCATION)
+
+    LaunchedEffect(cameraPermissionState.status) {
+        if (!cameraPermissionState.status.isGranted) {
+            cameraPermissionState.launchPermissionRequest()
+        }
+    }
+
+    val taiwan = LatLng(22.999973101427155, 120.21985214463398)
+    val cameraPositionState = rememberCameraPositionState {
+        position = CameraPosition.fromLatLngZoom(taiwan, 15f)
+    }
+
+    val bottomBarHeight = 80.dp // default NavigationBarHeight is 80.dp
+    val context = LocalContext.current
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(bottom = bottomBarHeight)
+    ) {
+        if (cameraPermissionState.status.isGranted) {
+            GoogleMap(
+                modifier = Modifier.fillMaxSize(),
+                cameraPositionState = cameraPositionState,
+                properties = MapProperties(isMyLocationEnabled = true), // 啟用當前位置
+                uiSettings = MapUiSettings(
+                    myLocationButtonEnabled = true // 啟用定位按鈕
+                )
+            ) {
+                Marker(
+                    state = MarkerState(position = taiwan),
+                    title = "library",
+                    snippet = "國立成功大學圖書館"
+                )
+            }
+        } else {
+//            Text(
+//                modifier = Modifier
+//                    .fillMaxSize()
+//                    .wrapContentSize(Alignment.Center),
+//                text = "No permission"
+//            )
+            Button(
+                onClick = {
+                    context.startActivity(Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).apply {
+                        data = Uri.fromParts("package", "com.example.lifelinealert", null)
+                    })
+                },
+                modifier = Modifier
+                    .fillMaxSize()
+                    .wrapContentSize(Alignment.Center),
+                colors = ButtonDefaults.buttonColors(containerColor = Color.Red)
+            ) {
+                Text("需要精準位置權限授權，點擊跳轉設定")
+            }
+        }
+    }
+}
