+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.8.0" # original: agp = "8.4.0"
+agp = "8.4.0" # original: agp = "8.4.0"
 kotlin = "1.9.0"
 coreKtx = "1.12.0" # original: coreKtx = "1.15.0"
 junit = "4.13.2"
